Created an etl_process folder inside pipeline to separate files in accordance
